+++ src/test/resources/tipitapi/drawmytoday/domain/adreward/repository/ExpiredAdReward.sql
