+++ src/test/resources/tipitapi/drawmytoday/domain/adreward/repository/ValidAdReward.sql
