+++ src/test/resources/tipitapi/drawmytoday/domain/ticket/repository/ValidTicket.sql
